% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_drivers}
\alias{calculate_drivers}
\title{Calculate drivers as defined in questions table.}
\usage{
calculate_drivers(x, questions = WESquestions18)
}
\arguments{
\item{x}{A data.frame with survey responses on a scale from 1 to 5.}

\item{questions}{A data.frame containing the questions used in the survey. Default is the questions file from WES 2018. See WESquestions18 for reference.}
}
\value{
The initial data.frame with additional columns for each driver as listed in the questions table.
}
\description{
Calculate drivers as defined in questions table.
}
