% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_engagement}
\alias{calculate_engagement}
\title{Calculate engagement as defined by user.}
\usage{
calculate_engagement(x, eng_char = c("BC Public Service Commitment",
  "Organization Satisfaction", "Job Satisfaction"))
}
\arguments{
\item{x}{A data.frame with survey responses on a scale from 1 to 5.}

\item{eng_char}{A character vector with the names of the columns of x used to calculate engagement. Default is c("BC Public Service Commitment", "Organization Satisfaction", "Job Satisfaction")}
}
\value{
The initial data.frame with additional column Engagement.
}
\description{
Calculate engagement as defined by user.
}
