% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_scores}
\alias{calculate_scores}
\title{Convert 1 - 5 scale to 100pt scale.}
\usage{
calculate_scores(x, questions = WESquestions18)
}
\arguments{
\item{x}{A data.frame with survey responses on a scale from 1 to 5.}

\item{questions}{A data.frame containing the questions used in the survey. Default is the questions file from WES 2018. See WESquestions18 for reference.}
}
\value{
The initial data.frame with additional columns for each 100pt scale recode. Additional column names have the format <Q_COLUMN>S.
}
\description{
Convert 1 - 5 scale to 100pt scale.
}
