% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{rounded}
\alias{rounded}
\title{rounded}
\usage{
rounded(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Rounds positive numbers to integers (i.e., zero decimal places).
}
\details{
The base R round function "rounds to even" when rounding off a 5 (e.g., 0.5 rounds to 0 and 1.5 rounds to 2). To ensure that 0.5 is rounded up,
\code{rounded} adds 0.5 to the input and truncates. Note that this function only works for positive numbers.
}
\examples{
rounded(.5 + 0:4) # 1  2  3  4  5
## different from IEEE / IEC rounding: 0  2  2  4  4
}
