% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_responserate_data}
\alias{tabulate_responserate_data}
\title{Tabulates response rates and logs to a csv file.}
\usage{
tabulate_responserate_data(x, reportID, reportTitle,
  responseFlag = "QCOMP", outputFile = "Response_Log.csv",
  userList = NULL)
}
\arguments{
\item{x}{A data.frame}

\item{reportID}{A character vector containing the column names in x containing the reporting IDs (e.g., ORGID, LEVEL1_ID, etc.)}

\item{reportTitle}{A character vector containing the column names in x containing the report titles (e.g., Organization, LEVEL1, etc.).}

\item{responseFlag}{A character string of the column in data flagging completions. Default is "QCOMP".}

\item{outputFile}{The path to the output file. Default is "Response_Log.csv"}
}
\value{
A data.frame with tabulated response rates
}
\description{
Tabulates response rates and logs to a csv file.
}
\examples{
tabulate_responserate_data(data,reportID="ID",reportTitle = "Title",outputFile="BCPS_Response_Log.csv")

}
