% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_responserate_demogdata}
\alias{tabulate_responserate_demogdata}
\title{Tabulates response rates and logs to a csv file - for demographic breakdowns.}
\usage{
tabulate_responserate_demogdata(x, reportID, demogRollup,
  responseFlag = "QCOMP", outputFile = "Response_Log.csv")
}
\arguments{
\item{x}{A data.frame}

\item{reportID}{A character vector containing the column names in x containing the reporting IDs (e.g., ORGID, LEVEL1_ID, etc.)}

\item{demogRollup}{A data.frame containing the demographic groupings to be reported. See WESdemogRollup18, or WESOrgdemogRollup18 for reference. Must include the columns reportID, Demographic, Group, Demographic_in_data, Group_in_data and Sort.}

\item{responseFlag}{A character string of the column in data flagging completions. Default is "QCOMP".}

\item{outputFile}{The path to the output file. Default is "Response_Log.csv"}
}
\value{
A data.frame with tabulated response rates
}
\description{
Tabulates response rates and logs to a csv file - for demographic breakdowns.
}
\examples{
tabulate_responserate_demogdata(x=data, reportID="ORGID18", demogRollup = demog_rollup_org)

}
