% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WESquestions18}
\alias{WESquestions18}
\title{Questions file used for 2018 WES.}
\format{A data frame with 81 rows and 7 variables:
\describe{
\item{Q_TEXT_FORMAT}{Same as Q_TEXT, unless additional latex formatting is desired for external reports (e.g., italics)}
}}
\usage{
WESquestions18
}
\description{
A dataset containing the QIDs, Model Linkage, Question text, Survey Sections and driver IDs
for each question in the 2018 WES.
}
\keyword{datasets}
