% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_responserate_sheet}
\alias{write_responserate_sheet}
\title{Create a formatted excel response rate sheet from tabulated data.}
\usage{
write_responserate_sheet(x, outputFile, inputFile = NULL,
  sheetName = "Response Rates")
}
\arguments{
\item{x}{A data.frame}

\item{outputFile}{The path to the output file.}

\item{inputFile}{The path to the input file if one exists. For example to append sheet to a readme.xlsx file. If NULL, a new workbook will be created.}

\item{sheetName}{A character string with the sheet name. The default is "Response Rates".}
}
\description{
Create a formatted excel response rate sheet from tabulated data.
}
