% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_scores}
\alias{calculate_scores}
\alias{calculate_categories}
\alias{calculate_drivers}
\alias{calculate_engagement}
\title{Calculate Survey Scores and Categories}
\usage{
calculate_scores(x, questions = WESquestions18)

calculate_categories(x, questions = WESquestions18)

calculate_drivers(x, questions = WESquestions18)

calculate_engagement(x, questions = WESquestions18,
  eng_char = c("BC Public Service Commitment",
  "Organization Satisfaction", "Job Satisfaction"))
}
\arguments{
\item{x}{A data.frame.}

\item{questions}{A data.frame containing the questions used in the survey. Default is the questions file from WES 2018. See \code{WESquestions18} for reference.}

\item{eng_char}{A character vector with the names of the \emph{driver} variables used to calculate \code{Engagement}.}
}
\value{
A data.frame with the same variables as x plus additional columns depending on the function used.
\emph{100pt scale} variables have the format <Q_COLUMN>S and \emph{category} variables have the format <Q_COLUMN>_3cat.
}
\description{
\code{calculate_scores} adds new \emph{100pt scale} variables based on the \emph{5pt scale} variables provided.

\code{calculate_categories}  adds new \emph{category} (i.e., \emph{3pt scale}) variables based on the \emph{5pt scale} variables provided.

\code{calculate_drivers} adds new \emph{driver} variables calculated as the average of the \emph{100pt scale} variables for that driver, defined by \code{questions$MODEL_LINKAGE}.

\code{calculate_engagement} adds a new \emph{engagement} variable calculated as the average of the \emph{driver} variables in \code{eng_char}.
}
\details{
\code{calculate_scores} takes a data.frame containing \emph{5pt scale} variables and appends new \emph{100pt scale} variables based on the following conversion:
\itemize{
\item 1 --> 0
\item 2 --> 25
\item 3 --> 50
\item 4 --> 75
\item 5 --> 100
}

\code{calculate_categories} takes a data.frame containing \emph{5pt scale} variables and appends new \emph{category} variables based on the following conversion:
\itemize{
\item 1 --> 1
\item 2 --> 1
\item 3 --> 2
\item 4 --> 3
\item 5 --> 3
}

The questions data.frame input provides the framework for the survey. It must contain the variables
\code{Q_ID,  Q_COLUMN, MODEL_LINKAGE, Q_TEXT, Q_TEXT_FORMAT, Q_SECTION,} and \code{Driver_ID}. \code{Q_COLUMN} contains the names of the \emph{5pt scale} variables of x.
\code{MODEL_LINKAGE} provides the bridge between the survey questions and the drivers in the WES model. See the default questions data.frame \code{WESquestions18} for details.
}
\note{
As the \emph{driver} variables are needed to calculate \code{Engagement}, running \code{calculate_engagement} will call \code{calculate_drivers}. Similarly,
as the \emph{100pt scale} variables are needed to calculate the \emph{drivers}, \code{calculate_drivers} will call \code{calculate_scores}.
Hence, a call to \code{calculate_engagement}, will result in the question and driver scores being calculated and appended to the returning data.frame.
}
