% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate_driverscore_demogdata}
\alias{tabulate_driverscore_demogdata}
\title{Tabulates driver/engagement scores and logs to a csv file - for demographic breakdowns.}
\usage{
tabulate_driverscore_demogdata(x, reportID, demogRollup,
  questions = WESquestions18, outputFile = "Drivers_Log.csv")
}
\arguments{
\item{x}{A data.frame}

\item{reportID}{A character vector containing the column names in x containing the reporting IDs (e.g., ORGID, LEVEL1_ID, etc.)}

\item{demogRollup}{A data.frame containing the demographic groupings to be reported. See WESdemogRollup18, or WESOrgdemogRollup18 for reference. Must include the columns reportID, Demographic, Group, Demographic_in_data, Group_in_data and Sort.}

\item{questions}{A data.frame containing the questions used in the survey. Default is the questions file from WES 2018. See WESquestions18 for reference.}

\item{outputFile}{The path to the output file. Default is "Driver_Log.csv"}
}
\value{
A data.frame with tabulated response rates
}
\description{
Tabulates driver/engagement scores and logs to a csv file - for demographic breakdowns.
}
\examples{
tabulate_driverscore_demogdata(data, reportID="ORGID18", demogRollup = demog_rollup_org)

}
